useInputValidation added and removes a lot of the boiler-plate validation logic. Is overridable e.g. ZipInput and BirthdateInput do this. Bump prerelease version and regenerate design-system.css
