+++ src/components/Inputs/CheckboxInput/CheckboxInput.module.scss
@@ -41,7 +41,7 @@
   background-color: var(--BrandForest);
   border-width: 0px;
   opacity: 1;
-  fill: var(--white)
+  fill: var(--white);
 }
 
 .CheckboxInput:focus ~ .Facade {
