+++ src/components/design-system.css
@@ -792,213 +792,6 @@ a,
 
 /* @getethos/design-system/Media.scss
    ========================================================================== */
-/*
-  Flexbox globals
-  Taken from:
-  https://github.com/tachyons-css/tachyons/blob/master/src/_flexbox.css
-
-  Don't re-import this in funnel! This is just here to make the 
-  universal navbar work in CMS.
-
-  This is imported globally in the funnel but not the cms yet.
-  Nick used some stuff here to write the universal navbar.
-*/
-.full-width {
-  width: 100%;
-}
-
-.full-height {
-  height: 100%;
-}
-
-.flex {
-  display: flex;
-}
-
-.flex-column {
-  flex-direction: column;
-}
-
-.flex-row {
-  flex-direction: row;
-}
-
-.flex-wrap {
-  flex-wrap: wrap;
-}
-
-.items-start {
-  align-items: flex-start;
-}
-
-.items-end {
-  align-items: flex-end;
-}
-
-.items-center {
-  align-items: center;
-}
-
-.justify-end {
-  justify-content: flex-end;
-}
-
-.justify-center {
-  justify-content: center;
-}
-
-.justify-between {
-  justify-content: space-between;
-}
-
-.justify-around {
-  justify-content: space-around;
-}
-
-.flex-shrink-0 {
-  flex-shrink: 0;
-}
-
-/*Namespace */
-.transforming-burger-button .iconWrapper {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  margin-bottom: 0;
-  cursor: pointer;
-  height: 16px;
-}
-
-.transforming-burger-button .icon {
-  --burger-width: 16px;
-  --burger-thickness: 2px;
-  cursor: pointer;
-  text-decoration: none;
-  border-bottom: none;
-}
-.transforming-burger-button .icon:hover,
-.transforming-burger-button .icon:focus,
-.transforming-burger-button .icon:active {
-  outline: none;
-}
-.transforming-burger-button .icon i {
-  position: relative;
-  z-index: 2;
-  display: block;
-  width: var(--burger-width);
-  height: var(--burger-thickness);
-  background: var(--GrayPrimary--translucent);
-  transition: all 0.2s ease-out;
-}
-.transforming-burger-button .icon i::before,
-.transforming-burger-button .icon i::after {
-  content: '';
-  width: var(--burger-width);
-  height: var(--burger-thickness);
-  background: var(--GrayPrimary--translucent);
-  position: absolute;
-  left: 0;
-  transition: 0.2s ease;
-}
-.transforming-burger-button .icon i::before {
-  top: -4px;
-}
-.transforming-burger-button .icon i::after {
-  bottom: -4px;
-}
-
-.transforming-burger-button .showMobileMenu i {
-  background: transparent;
-}
-.transforming-burger-button .showMobileMenu i::before {
-  top: 0;
-  transform: rotateZ(45deg);
-  background: white;
-}
-.transforming-burger-button .showMobileMenu i::after {
-  bottom: 0;
-  transform: rotateZ(-45deg);
-  background: white;
-}
-
-.UniversalNavbar--hamburger {
-  position: fixed;
-  top: 24px;
-  right: 24px;
-  z-index: 2;
-}
-@media (min-width: 600px) {
-  .UniversalNavbar--hamburger {
-    display: none;
-  }
-}
-
-.UniversalNavbar--phoneOnly {
-  height: var(--Space-64);
-  background-color: white;
-  padding: var(--Space-24) var(--ContainerPadding--Phone);
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
-@media (min-width: 600px) {
-  .UniversalNavbar--phoneOnly {
-    display: none;
-  }
-}
-
-.UniversalNavbar--tabletAndUp {
-  height: var(--Space-64);
-  background-color: white;
-  display: none;
-}
-@media (min-width: 600px) {
-  .UniversalNavbar--tabletAndUp {
-    display: block;
-  }
-}
-
-.UniversalNavbar--laptopAndUp {
-  display: none;
-}
-@media (min-width: 900px) {
-  .UniversalNavbar--laptopAndUp {
-    display: block;
-  }
-}
-
-.UniversalNavbar__tabletAndUpContainer {
-  max-width: 1440px;
-  margin: 0 auto;
-  padding: var(--Space-24);
-  height: var(--Space-64);
-  display: flex;
-  justify-content: space-between;
-}
-@media (min-width: 600px) and (max-width: 899px) {
-  .UniversalNavbar__tabletAndUpContainer {
-    padding: var(--Space-24) var(--ContainerPadding--Tablet);
-  }
-}
-@media (min-width: 900px) and (max-width: 1199px) {
-  .UniversalNavbar__tabletAndUpContainer {
-    padding: var(--Space-24) var(--ContainerPadding--Laptop);
-  }
-}
-@media (min-width: 1200px) {
-  .UniversalNavbar__tabletAndUpContainer {
-    padding: var(--Space-24) var(--ContainerPadding--Desktop);
-  }
-}
-
-.universal-navbar-logo {
-  height: 16px;
-}
-
-.universal-navbar-paddingLeft {
-  padding-left: 24px;
-}
-
 /**
    * Region:
    * Fancy animated scroll behavior
@@ -1069,102 +862,6 @@ a,
 }
 
 /* End region */
-/**
-   * Region:
-   * Big green mobile hamburger menu
-   */
-.shownMobileMenu,
-.hiddenMobileMenu {
-  position: fixed;
-  padding: 24px 24px 40px;
-  height: 100vh;
-  width: 100vw;
-  top: 0;
-  left: 0;
-  background-color: var(--BrandForest);
-  color: white;
-  z-index: 1;
-}
-.shownMobileMenu a,
-.shownMobileMenu a:visited,
-.shownMobileMenu a:hover,
-.hiddenMobileMenu a,
-.hiddenMobileMenu a:visited,
-.hiddenMobileMenu a:hover {
-  /* TODO REDESIGN: this should be a DS component */
-  font-weight: 300;
-  color: white;
-  text-decoration: unset !important;
-}
-
-.hiddenMobileMenu {
-  opacity: 0;
-  pointer-events: none;
-}
-
-/* End region */
-/*Namespace */
-.transforming-burger-button .iconWrapper {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  margin-bottom: 0;
-  cursor: pointer;
-  height: 16px;
-}
-
-.transforming-burger-button .icon {
-  --burger-width: 16px;
-  --burger-thickness: 2px;
-  cursor: pointer;
-  text-decoration: none;
-  border-bottom: none;
-}
-.transforming-burger-button .icon:hover,
-.transforming-burger-button .icon:focus,
-.transforming-burger-button .icon:active {
-  outline: none;
-}
-.transforming-burger-button .icon i {
-  position: relative;
-  z-index: 2;
-  display: block;
-  width: var(--burger-width);
-  height: var(--burger-thickness);
-  background: var(--GrayPrimary--translucent);
-  transition: all 0.2s ease-out;
-}
-.transforming-burger-button .icon i::before,
-.transforming-burger-button .icon i::after {
-  content: '';
-  width: var(--burger-width);
-  height: var(--burger-thickness);
-  background: var(--GrayPrimary--translucent);
-  position: absolute;
-  left: 0;
-  transition: 0.2s ease;
-}
-.transforming-burger-button .icon i::before {
-  top: -4px;
-}
-.transforming-burger-button .icon i::after {
-  bottom: -4px;
-}
-
-.transforming-burger-button .showMobileMenu i {
-  background: transparent;
-}
-.transforming-burger-button .showMobileMenu i::before {
-  top: 0;
-  transform: rotateZ(45deg);
-  background: white;
-}
-.transforming-burger-button .showMobileMenu i::after {
-  bottom: 0;
-  transform: rotateZ(-45deg);
-  background: white;
-}
-
 /*
   Flexbox globals
   Taken from:
